Attack/DexWave/requirements.txt
Yapsy==1.12.2
colorlog==4.2.1
tensorflow==2.3.0
python-dotenv==0.14.0